+++ CHANGELOG.md
@@ -23,6 +23,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
     - *OutputNode* finds nodes that don't have downstream nodes.
 - Traversal operations for geometry, providing a cleaner interface for
   operating on geometries.
+- GeoJSON Reader.
 
 ### Changed